shred: add 4K data alignment
This commit allows aligning output data by 4K to better match GNU shred.

The 4K block size is configured as a constant because it provides a
widely used value in the simplest way. However, there is a chance that
some systems may use a different value. So far, I haven't encountered
anything other than 4K, I decided not to overcomplicate the approach for
now.